support storing the db passphrase in an external keyring
Introduce the `org-roam-store-db-passphrase' setting to control the lifecycle of
the database encryption passphrase (used when `org-roam-encrypt-db' is set to
`sqlcipher'). By default, the passphrase is immediately discarded after each
use. This means the user has to type if fairly regularly (especially because
`org-roam-db-build-cache' forces a reconnect).

When setting `org-roam-store-db-passphrase' to `secret-service', the passphrase
will now be stored in an external keyring for the duration of the Emacs session
using the Emacs Secret Service API. This way the user only has to type it once
after start-up.