(feat)db: add support for encrypting the db
Add a new option to `org-roam-database-connector' that lets the user use
sqlcipher as the database backend. When that option is chosen, the user is
prompted for a passphrase at database opening time, which is used to
transparently encrypt the data on disk.

Also introduce the `org-roam-store-db-passphrase' setting to control the
lifecycle of the database encryption passphrase. By default, the passphrase is
immediately discarded after each use. When instead setting
`org-roam-store-db-passphrase' to `secret-service', the passphrase will be
stored in an external keyring for the duration of the Emacs session using the
Emacs Secret Service API.

The implementation could do with some factoring out, but the changes are
relatively self-contained.